+++ DU-Industry-Tool/changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v0.7.4
+
+- Industry attribute is now a clickable link to show all products producable on this type.
+- Recipe Search box now features auto-completion with suggestions.
+- Fixed doofus interim bug, where parts suddenly got schematics assigned.
+- Fixed tree group assignment for antimatter cores.
+- Fixes for industry assignments where the tier and size could be wrong.
+- Values in Schematics form are now read-only.
+- Minor layout changes to have ingredients' and products' columns have the same size.
+- Changed default font and switched internal scaling functionalities.
+
 ## v0.7.3
 
 - Added display of Industry below Products with tier and size.